backend_gradio/requirements.txt
amqp==5.1.0
async-timeout==4.0.2
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.6
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
FastAPI==0.95.2
Flask==2.1.1
Flask-Cors==3.0.10
flower==1.0.0
gunicorn==20.1.0
humanize==4.0.0
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
kombu==5.2.4
MarkupSafe==2.1.1
openai==0.27.2
packaging==21.3
postmark==1.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
PyGithub==1.58.1
pyparsing==3.0.8
pytz==2022.1
redis==4.2.2
six==1.16.0
tornado==6.1
typing-extensions==4.1.1
uvicorn==0.22.0
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.1.1
wrapt==1.14.0
zipp==3.8.0
urllib3==1.26.6
gradio<=3.32.0
python-dotenv==0.19.1

